+++ CHANGELOG.md
@@ -15,6 +15,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - missing unit tests
 - possibility to deactivate api request caching
+- logging for project controller
 
 ## [1.1.0](https://github.com/jazzica/wechange/compare/1.0.1...1.1.0) - 2022-11-03